app/requirements.txt
fastapi
uvicorn
python-multipart
torch
natsort
mysql-connector-python
python-dotenv
psycopg2-binary